fix #31 Deal with usage 'usageLimits' errors
GDriverRemoteRepository now depends on new GDriverClient interface,
which implements basic operations needed by the application,
such as 'deleteFolder' or 'createTopLevelFolder'.

The interface is implemented by GDriveDirectClient and GDriveRetryingClient,
where the latter serves as a proxy to the former, ensuring that all calls
to GDrive API are retried with exponential backoff, should any fail due
to usage rate limits being exceeded.